refactor: change UpdateTrackDTO filename
